api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
passlib[bcrypt]
PyJWT
prisma
